# src/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Basic Backend Interview Test
  description: Test
# during dev, should point to your local machine
host: localhost
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /:
    # binds a127 app logic to a route
    x-swagger-router-controller: 'hello.controller'
    get:
      description: Returns hello world! to the caller
      # used as the method name of the controller
      operationId: hello
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/HelloResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /neo/hazardous:
    # binds a127 app logic to a route
    x-swagger-router-controller: 'neo.controller'
    get:
      description: Returns all NEOs hazardous to the caller
      # used as the method name of the controller
      operationId: getNeosHazardous
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/NeosResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /neo/fastest:
    # binds a127 app logic to a route
    x-swagger-router-controller: 'neo.controller'
    get:
      description: Returns all NEOs hazardous to the caller
      # used as the method name of the controller
      operationId: getFastestNeo
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Neo"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
      parameters:
        - name: hazardous
          in: query
          description: The hazardous of the NEO to find
          required: false
          type: string
  /neo/best-year:
    # binds a127 app logic to a route
    x-swagger-router-controller: 'neo.controller'
    get:
      description: Returns all NEOs hazardous to the caller
      # used as the method name of the controller
      operationId: getBestYearNeo
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/BestYearResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
      parameters:
        - name: hazardous
          in: query
          description: The hazardous of the NEO to find
          required: false
          type: string
  /neo/best-month:
    # binds a127 app logic to a route
    x-swagger-router-controller: 'neo.controller'
    get:
      description: Returns all NEOs hazardous to the caller
      # used as the method name of the controller
      operationId: getBestMonthNeo
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/BestMonthResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
      parameters:
        - name: hazardous
          in: query
          description: The hazardous of the NEO to find
          required: false
          type: string
  /getLast3DaysNeosFromNasa:
    # binds a127 app logic to a route
    x-swagger-router-controller: 'neo.controller'
    get:
      description: Returns lasts 3 days NEOs got from Nasa to the caller
      # used as the method name of the controller
      operationId: getLast3DaysNeosFromNasa
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Neos"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:  
  HelloResponse:
    required:
      - hello
    properties:
      hello:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  BestMonthResponse:
    required:
      - month
      - count
    properties:
      month:
        type: string
      count:
        type: integer
  BestYearResponse:
    required:
      - year
      - count
    properties:
      year:
        type: string
      count:
        type: integer
  NeosResponse:  
    type: object
    properties:  
      neos:
        type: integer
        description: the number of neos in response
      total:
        type: integer
        description: the number of neos in all DB dataset
      _links:
        type: array
        items:
          $ref: '#/definitions/HALLink'         
      _embedded:
        type: array
        items:
          $ref: '#/definitions/Neo'
  Neos:  
    type: array
    items:
      $ref: '#/definitions/Neo'
  Neo:
    type: object
    properties:
      _links:
        type: array
        items:
          $ref: '#/definitions/HALLink'
      _id:
        type: integer
        description: The neo ID.
      reference:
        type: string
        description: reference of the NEO from NASA
      name:
        type: string
        description: name of the NEO
      speed:
        type: integer
        description: speed of the NEO in kilometers
      date:
        type: string
        format: date
        description: date of occurrency of the NEO from NASA
      hazardous:
        type: boolean
        description: If NEO is hazardous
      createdAt:
        type: string
        format: date
        description: date of NEO created in DB
      updatedAt:
        type: string
        format: date
        description: date of NEO updated in DB
  HALLink:
    type: object
    description: the HAL link by HATEOAS approach
    properties:
      href:
        type: string
        description: the full reference to a related resource
      rel:
        type: string  
        description: a reference to a related resource as secondary identifier